requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.7.0
django-extensions==4.1

channels==4.3.2
daphne==4.2.1
channels-redis==4.3.0

django-simple-captcha==0.6.2
django-ranged-response==0.2.0

psycopg2-binary==2.9.9
Pillow==12.0.0
bleach==6.3.0
asgiref==3.11.0
sqlparse==0.5.4
tzdata==2025.2
webencodings==0.5.1

celery==5.4.0
kombu==5.3.7
redis==5.1.1

django-elasticsearch-dsl==8.0
elasticsearch==8.15.1
elasticsearch-dsl==8.15.4
djangorestframework-simplejwt==5.3.1

